reorganised atomic design components
