+++ src/components/Pages/Group/Group.tsx
@@ -8,7 +8,7 @@ import { IGroup } from "../../../types/databaseTypes";
 // Components
 import Loader from "../../Atoms/Loader/Loader";
 import GroupDetailsForm from "../../Templates/DetailsForm/GroupDetailsForm";
-import ErrorBox from "../../Molecules/ErrorBox/ErrorBox";
+import ErrorBox from "../../Atoms/ErrorBox/ErrorBox";
 // Supabase
 import { readGroupById } from "../../../../supabase/groupsFunctions";
